feat(01-security-hardening-04): add comprehensive security unit tests
- test_node_identity.cpp: 9 tests — deterministic sigs, sign/verify roundtrip,
  AES-256-GCM encrypted save/load, wrong passphrase, tampered file, edge cases
- test_message_signing.cpp: 7 tests — valid verify, tampered payload, wrong key,
  empty/truncated sig, replay protection, expired timestamp rejection
- test/CMakeLists.txt: registered test_node_identity and test_message_signing targets
- Total: 16 Google Test cases covering SEC-01 through SEC-06
