--- test/security/test_security.cpp
@@ -1,364 +1,0 @@
-/**
- * @file       test_security.cpp
- * @brief      Unit tests for NodeIdentity and MessageSigning (TDD for plan 01-02)
- * @date       2026-05-28
- * @author     GSD Executor
- */
-
-#include <gtest/gtest.h>
-#include "security/NodeIdentity.hpp"
-#include "security/MessageSigning.hpp"
-
-#include <cstdio>
-#include <fstream>
-#include <iomanip>
-#include <sstream>
-#include <string>
-#include <vector>
-
-using namespace sgns::neoswarm;
-using namespace sgns::neoswarm::security;
-
-// =======================================================================
-// Task 1: RFC6979 Deterministic Nonces
-// =======================================================================
-
-TEST( NodeIdentity, DeterministicSignature )
-{
-    // Signing the same message twice with the same key MUST produce
-    // identical DER-encoded signatures (deterministic nonces).
-    NodeIdentity ident;
-    ASSERT_TRUE( ident.Generate().has_value() );
-    ASSERT_TRUE( ident.IsLoaded() );
-
-    std::vector<uint8_t> msg1 = {0x01, 0x02, 0x03, 0x04};
-    std::vector<uint8_t> msg2 = {0x01, 0x02, 0x03, 0x04};  // same message
-
-    auto sig1 = ident.Sign( msg1 );
-    auto sig2 = ident.Sign( msg2 );
-    ASSERT_TRUE( sig1.has_value() );
-    ASSERT_TRUE( sig2.has_value() );
-
-    EXPECT_EQ( sig1.value().size(), sig2.value().size() );
-    EXPECT_EQ( sig1.value(), sig2.value() );
-}
-
-TEST( NodeIdentity, DifferentMessagesDifferentSignatures )
-{
-    // Signing different messages MUST produce different signatures.
-    NodeIdentity ident;
-    ASSERT_TRUE( ident.Generate().has_value() );
-
-    std::vector<uint8_t> msgA = {0xAA};
-    std::vector<uint8_t> msgB = {0xBB};
-
-    auto sigA = ident.Sign( msgA );
-    auto sigB = ident.Sign( msgB );
-    ASSERT_TRUE( sigA.has_value() );
-    ASSERT_TRUE( sigB.has_value() );
-
-    EXPECT_NE( sigA.value(), sigB.value() );
-}
-
-TEST( NodeIdentity, SignAndVerifyRoundtrip )
-{
-    // The produced signature MUST verify against the node's public key.
-    NodeIdentity ident;
-    ASSERT_TRUE( ident.Generate().has_value() );
-
-    std::vector<uint8_t> msg = {0x01, 0x02, 0x03, 0x04, 0x05};
-    auto sig = ident.Sign( msg );
-    ASSERT_TRUE( sig.has_value() );
-
-    EXPECT_TRUE( ident.Verify( msg, sig.value() ) );
-}
-
-// =======================================================================
-// Task 2: Real MessageSigning::Verify
-// =======================================================================
-
-namespace
-{
-    std::string PubKeyToHex( const NodeIdentity::PubKey &key )
-    {
-        std::ostringstream oss;
-        for ( auto b : key )
-        {
-            oss << std::hex << std::setw( 2 ) << std::setfill( '0' )
-                << static_cast<int>( b );
-        }
-        return oss.str();
-    }
-}
-
-TEST( MessageSigning, VerifyValidSignature )
-{
-    // Sign a message with NodeIdentity, verify with MessageSigning::Verify
-    // using the correct pub_key_hex — MUST return true.
-    NodeIdentity ident;
-    ASSERT_TRUE( ident.Generate().has_value() );
-
-    const std::string pubKeyHex = PubKeyToHex( ident.PublicKey() );
-    MessageSigning   signer( ident );
-
-    const std::string payload = R"({"msg":"hello"})";
-    auto              sig     = signer.Sign( payload );
-    ASSERT_TRUE( sig.has_value() );
-    ASSERT_FALSE( sig.value().empty() );
-
-    EXPECT_TRUE( MessageSigning::Verify( payload, sig.value(), pubKeyHex ) );
-}
-
-TEST( MessageSigning, VerifyTamperedPayload )
-{
-    // Tamper with the payload after signing, verify — MUST return false.
-    NodeIdentity ident;
-    ASSERT_TRUE( ident.Generate().has_value() );
-
-    const std::string pubKeyHex = PubKeyToHex( ident.PublicKey() );
-    MessageSigning   signer( ident );
-
-    const std::string payload = R"({"msg":"hello"})";
-    auto              sig     = signer.Sign( payload );
-    ASSERT_TRUE( sig.has_value() );
-
-    // Tamper: change the payload
-    const std::string tampered = R"({"msg":"world"})";
-    EXPECT_FALSE( MessageSigning::Verify( tampered, sig.value(), pubKeyHex ) );
-}
-
-TEST( MessageSigning, VerifyWrongKey )
-{
-    // Use a different node's public key to verify — MUST return false.
-    NodeIdentity identA;
-    NodeIdentity identB;
-    ASSERT_TRUE( identA.Generate().has_value() );
-    ASSERT_TRUE( identB.Generate().has_value() );
-
-    const std::string pubKeyB = PubKeyToHex( identB.PublicKey() );
-    MessageSigning   signerA( identA );
-
-    const std::string payload = R"({"msg":"hello"})";
-    auto              sig     = signerA.Sign( payload );
-    ASSERT_TRUE( sig.has_value() );
-
-    // Verify identA's signature with identB's key — should fail
-    EXPECT_FALSE( MessageSigning::Verify( payload, sig.value(), pubKeyB ) );
-}
-
-TEST( MessageSigning, VerifyEmptySignature )
-{
-    // Provide an empty signature — MUST return false.
-    NodeIdentity ident;
-    ASSERT_TRUE( ident.Generate().has_value() );
-
-    const std::string pubKeyHex = PubKeyToHex( ident.PublicKey() );
-
-    EXPECT_FALSE( MessageSigning::Verify( "payload", {}, pubKeyHex ) );
-}
-
-TEST( MessageSigning, VerifyTruncatedSignature )
-{
-    // Provide a truncated/corrupted DER signature — MUST return false.
-    NodeIdentity ident;
-    ASSERT_TRUE( ident.Generate().has_value() );
-
-    const std::string pubKeyHex = PubKeyToHex( ident.PublicKey() );
-
-    // Truncated DER: just 8 bytes
-    std::vector<uint8_t> truncated = {0x30, 0x06, 0x02, 0x01, 0x01, 0x02, 0x01, 0x01};
-    EXPECT_FALSE( MessageSigning::Verify( "payload", truncated, pubKeyHex ) );
-}
-
-// =======================================================================
-// Task 3: Nonce + Timestamp Replay Protection
-// =======================================================================
-
-TEST( MessageSigning, VerifyAndStripValid )
-{
-    // A freshly signed message verifies successfully and
-    // VerifyAndStrip extracts the original payload.
-    NodeIdentity ident;
-    ASSERT_TRUE( ident.Generate().has_value() );
-
-    const std::string pubKeyHex = PubKeyToHex( ident.PublicKey() );
-    MessageSigning   signer( ident );
-
-    const std::string original = R"({"msg":"hello"})";
-    std::string       payload  = signer.AttachSignature( original );
-
-    // Payload should differ from original (signature field appended)
-    EXPECT_NE( payload, original );
-
-    // VerifyAndStrip should return true and restore original payload
-    EXPECT_TRUE( MessageSigning::VerifyAndStrip( payload, pubKeyHex ) );
-    EXPECT_EQ( payload, original );
-}
-
-TEST( MessageSigning, VerifyAndStripExpiredTimestamp )
-{
-    // A message with a timestamp older than the replay window
-    // (30s) is rejected.
-    NodeIdentity ident;
-    ASSERT_TRUE( ident.Generate().has_value() );
-
-    const std::string pubKeyHex = PubKeyToHex( ident.PublicKey() );
-    MessageSigning   signer( ident );
-
-    // Build a signed payload then inject an expired timestamp
-    std::string payload = signer.AttachSignature( R"({"msg":"test"})" );
-
-    // Find and replace the "ts" field with an old timestamp
-    auto tsPos = payload.rfind( ",\"ts\":" );
-    ASSERT_NE( tsPos, std::string::npos );
-
-    // Replace timestamp with one from 60 seconds ago
-    uint64_t oldTs = MessageSigning::CurrentTimestampMs() - 61000;
-    auto     tsVal = payload.find_first_of( "0123456789", tsPos + 6 );
-    auto     tsEnd = payload.find_first_of( ",}", tsVal );
-    ASSERT_NE( tsVal, std::string::npos );
-    ASSERT_NE( tsEnd, std::string::npos );
-
-    payload.replace( tsVal, tsEnd - tsVal, std::to_string( oldTs ) );
-
-    // VerifyAndStrip should reject the expired message
-    EXPECT_FALSE( MessageSigning::VerifyAndStrip( payload, pubKeyHex ) );
-}
-
-// =======================================================================
-// Task: AES-256-GCM Encrypted Key Storage (TDD tests)
-// =======================================================================
-
-namespace
-{
-    const std::string kTestKeyPath  = "/tmp/gnus_test_node.key";
-    const std::string kTestPass     = "test123";
-    const std::string kWrongPass    = "wrong456";
-
-    void RemoveTestFile()
-    {
-        std::remove( kTestKeyPath.c_str() );
-    }
-} // namespace
-
-TEST( NodeIdentity, SaveEncryptedLoadEncryptedRoundtrip )
-{
-    // Test 1: Generate a keypair, SaveEncrypted with "test123",
-    // LoadEncrypted with same passphrase → PeerId identical.
-    RemoveTestFile();
-
-    NodeIdentity ident1;
-    ASSERT_TRUE( ident1.Generate().has_value() );
-    ASSERT_TRUE( ident1.IsLoaded() );
-
-    auto saveResult = ident1.SaveEncrypted( kTestKeyPath, kTestPass );
-    ASSERT_TRUE( saveResult.has_value() );
-
-    NodeIdentity ident2;
-    auto         loadResult = ident2.LoadEncrypted( kTestKeyPath, kTestPass );
-    ASSERT_TRUE( loadResult.has_value() );
-    ASSERT_TRUE( ident2.IsLoaded() );
-
-    EXPECT_EQ( ident1.PeerId(), ident2.PeerId() );
-
-    RemoveTestFile();
-}
-
-TEST( NodeIdentity, LoadEncryptedWrongPassphrase )
-{
-    // Test 2: LoadEncrypted with wrong passphrase → IdentityError.
-    RemoveTestFile();
-
-    NodeIdentity ident1;
-    ASSERT_TRUE( ident1.Generate().has_value() );
-    ASSERT_TRUE( ident1.SaveEncrypted( kTestKeyPath, kTestPass ).has_value() );
-
-    NodeIdentity ident2;
-    auto         result = ident2.LoadEncrypted( kTestKeyPath, kWrongPass );
-
-    // With stub: returns IdentityError. With real impl: GCM tag fails.
-    EXPECT_FALSE( result.has_value() );
-    EXPECT_EQ( result.error(), Error::IdentityError );
-
-    RemoveTestFile();
-}
-
-TEST( NodeIdentity, LoadEncryptedTamperedFile )
-{
-    // Test 3: SaveEncrypted, tamper one byte of file, LoadEncrypted → IdentityError.
-    RemoveTestFile();
-
-    NodeIdentity ident1;
-    ASSERT_TRUE( ident1.Generate().has_value() );
-    ASSERT_TRUE( ident1.SaveEncrypted( kTestKeyPath, kTestPass ).has_value() );
-
-    // Tamper with one byte in the file (past the header: salt_len + salt + IV)
-    {
-        std::fstream f( kTestKeyPath, std::ios::binary | std::ios::in | std::ios::out );
-        ASSERT_TRUE( f.is_open() );
-
-        // Seek past: 4B salt_len + 32B salt + 12B IV = 48 bytes, tamper the first ciphertext byte
-        f.seekp( 48, std::ios::beg );
-        char c = 0;
-        f.get( c );
-        f.seekp( 48, std::ios::beg );
-        f.put( static_cast<char>( c ^ 0xFF ) );  // flip bits
-        f.close();
-    }
-
-    NodeIdentity ident2;
-    auto         result = ident2.LoadEncrypted( kTestKeyPath, kTestPass );
-
-    EXPECT_FALSE( result.has_value() );
-    EXPECT_EQ( result.error(), Error::IdentityError );
-
-    RemoveTestFile();
-}
-
-TEST( NodeIdentity, SaveEncryptedWithoutKey )
-{
-    // Test 4: SaveEncrypted without a key loaded → IdentityError.
-    RemoveTestFile();
-
-    NodeIdentity ident;
-    ASSERT_FALSE( ident.IsLoaded() );
-
-    auto result = ident.SaveEncrypted( kTestKeyPath, kTestPass );
-
-    EXPECT_FALSE( result.has_value() );
-    EXPECT_EQ( result.error(), Error::IdentityError );
-
-    RemoveTestFile();
-}
-
-TEST( NodeIdentity, LoadEncryptedNonexistentFile )
-{
-    // Test 5: LoadEncrypted on a non-existent file → IdentityError.
-    RemoveTestFile();
-
-    NodeIdentity ident;
-    auto         result = ident.LoadEncrypted( kTestKeyPath, kTestPass );
-
-    EXPECT_FALSE( result.has_value() );
-    EXPECT_EQ( result.error(), Error::IdentityError );
-}
-
-TEST( NodeIdentity, SaveEncryptedOverwrite )
-{
-    // Test 6: SaveEncrypted then SaveEncrypted again → file overwritten (second call succeeds).
-    RemoveTestFile();
-
-    NodeIdentity ident1;
-    ASSERT_TRUE( ident1.Generate().has_value() );
-    ASSERT_TRUE( ident1.SaveEncrypted( kTestKeyPath, kTestPass ).has_value() );
-
-    // Second save with same key → should succeed (overwrite)
-    ASSERT_TRUE( ident1.SaveEncrypted( kTestKeyPath, kTestPass ).has_value() );
-
-    // Verify file is still loadable
-    NodeIdentity ident2;
-    ASSERT_TRUE( ident2.LoadEncrypted( kTestKeyPath, kTestPass ).has_value() );
-    EXPECT_EQ( ident1.PeerId(), ident2.PeerId() );
-
-    RemoveTestFile();
-}
